Sort the players by score.
The reasoning that this is done in the behavior is to keep the rendering
simple and not intertwined with application logic. Also, this method
ensures that the list is only changed when the score sort order changes.
(Otherwise the rendering would either have to figure out whether the
order changed or always sort&redraw.)